Update README.md
Just read "how google dealt with mobile" - that was a heafty article! not entirley about mobile, more a quick summery of the big leaps in tech and interaction that they have made. Brings up some interesting ethical thoughts as well regarding the breadth & depth of knowledge on individuals and the population.